Refactor protocol modules
