+++ src/io/sync.rs
@@ -27,8 +27,9 @@ use {
         config::Config,
         error::{ClientResult, ConnectionSetupError, Error},
         protocol::{
-            ClientHandshake, DecodeState, Decoder, MRespState, PipelineResult, RState,
-            ServerHandshake,
+            handshake::{ClientHandshake, ServerHandshake},
+            state_init::{DecodeState, MRespState, PipelineResult, RState},
+            Decoder,
         },
         query::Pipeline,
         response::{FromResponse, Response},
